Example of web flutter driver
