+++ packages/flutter_driver/lib/src/extension/io_extension.dart
@@ -1,0 +1,5 @@
+
+
+void registerWebServiceExtension(Future<Map<String, dynamic>> Function(Map<String, String>) call) {
+  throw UnsupportedError('Use registerServiceExtension instead');
+}
