+++ packages/flutter_driver/pubspec.yaml
@@ -14,6 +14,8 @@ dependencies:
   path: 1.6.4
   web_socket_channel: 1.0.15
   vm_service_client: 0.2.6+2
+  webkit_inspection_protocol:
+  js:
   flutter:
     sdk: flutter
   flutter_test:
